+++ js/util.js
@@ -1,0 +1,19 @@
+
+function getRandomPositiveInteger(a, b) {
+
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+function checkStringLength(string, length) {
+  return string.length <= length;
+}
+
+export{getRandomPositiveInteger, getRandomArrayElement, checkStringLength};
+
